Update CT 7
I had to make changes to my code because it's not letting me running and give me some strange error code but after I chnaged it still giving me the same error codes. So i'll tinker with it and see if i can fix it but i'll include this as my 2nd edit.