Update CT 7
Okay well it turned out the previous code I made worked. and this code worked. I forgot to name my file with the .cpp and the command to run it was also wrong. Once I fixed that it ran for both of them. so i'm using my original code since it's a lot simpler at least to me. but yeah no changes should have been made. It worked but I made some mistakes and overlooked some basic methods.